Traceback (most recent call last):   File "/content/drive/My Drive/UNet-Pytorch/train.py", line 138, in <module>     loss = loss_fn(preds, masks)   File "/usr/local/lib/python3.6/dist-packages/torch/nn/modules/module.py", line 541, in __call__     result = self.forward(*input, **kwargs)   File "/usr/local/lib/python3.6/dist-packages/torch/nn/modules/loss.py", line 916, in forward     ignore_index=self.ignore_index, reduction=self.reduction)   File "/usr/local/lib/python3.6/dist-packages/torch/nn/functional.py", line 2009, in cross_entropy     return nll_loss(log_softmax(input, 1), target, weight, None, ignore_index, None, reduction)   File "/usr/local/lib/python3.6/dist-packages/torch/nn/functional.py", line 1840, in nll_loss     ret = torch._C._nn.nll_loss2d(input, target, weight, _Reduction.get_enum(reduction), ignore_index) RuntimeError: Expected object of scalar type Long but got scalar type Float for argument #2 'target' in call to _thnn_nll_loss2d_forward
Details:
  change mask to long type tensor

+++ transforms.py
@@ -225,8 +225,7 @@ class ToTensor:
         img = torch.from_numpy(img)
         img = img.float() / 255.0
 
-        mask = torch.from_numpy(mask)
-        mask = mask.float()
+        mask = torch.from_numpy(mask).long()
 
         return img, mask
